mtk std libary
