+++ .experiment/ontology.jl
@@ -1,8 +1,6 @@
 # see Cobos Mendez et al. (2020), Fig. 5, Table 3, and Table 4
 using Graphs, MetaGraphsNext, Symbolics, ModelingToolkit
-
-include("energypair.jl")
-include("ports.jl")
+using ModelingToolkit: t_nounits as t, D_nounits as D
 
 # Maybe 'component' should be 'element' which includes functionality for
 # components and junctions. Then dispatch on the parametric type:
@@ -22,6 +20,10 @@ include("ports.jl")
 # these define a fixed definition that is reusable across a bond graph definition
 # then the Component struct is specifically for repeated *instances* within a single bond graph model
 # TODO See https://docs.julialang.org/en/v1/manual/methods/#Function-like-objects
+
+# TODO I should think of the model-building side of this package as generating MTK models
+# from a graph description or interface - I don't need to reinvent the wheel when it comes
+# to definining ports and components
 
 abstract type BondGraphVertexClass end
 
@@ -54,14 +56,16 @@ abstract type EqualFlow <: NonParametricJunction end
 
 ############################################################################
 
-abstract type AbstractNode end  # with the new ontology, this may not be needed?
+# abstract type AbstractNode end  # with the new ontology, this may not be needed?
 
-struct Component{T<:BondGraphVertexClass} <: AbstractNode
+# FIXME CONTINUE FROM HERE
+# Challenge: how to test that MTK models are chosen correctly
+# i.e. that a storage component is really a storage component
+
+struct Component{T<:BondGraphVertexClass}
+    class::T
     name::Symbol
-    variables::Vector{Num}
-    parameters::Vector{Num}
-    equations::Vector{Equation}
-    ports::Vector{Port}
+    model::ModelingToolkit.AbstractSystem
 end
 function Component{T}(; name, variables=Num[], parameters=Num[], equations=Equation[], ports=Port[]) where {T<:BondGraphVertexClass}
     # empty 'generic' component
@@ -121,94 +125,7 @@ function nextfreeport(junc::Component{<:NonParametricJunction})
     port
 end
 
-struct Bond
-    src::Port
-    dst::Port
-    function Bond(src::Port, dst::Port)
-        src.connected = true
-        dst.connected = true
-        new(src, dst)
-    end
-end
-function Bond(src_comp::Component, dst_comp::Component)
-    hasfreeport(src_comp) || error("$src_comp has no free ports")
-    hasfreeport(dst_comp) || error("$dst_comp has no free ports")
-    Bond(nextfreeport(src_comp), nextfreeport(dst_comp))
-end
-
-# show(io::IO, b::Bond) = print(io, "Bond()")  # FIXME
-
-# is_connected.(capacitor.ports)
-
-# hasfreeport(capacitor)
-# hasfreeport(kvl)
-
-# nextfreeport(capacitor)
-# nextfreeport(kvl)
-
 ############################################################################
-import Base: size
-
-# New bond graph structure
-mutable struct NewBondGraph <: AbstractGraph{Int64}
-    name::Symbol
-    graph::MetaGraph
-    function NewBondGraph(graph::AbstractGraph; name::Symbol)
-        # creating MetaGraph in a constructor keeps it type stable
-        metagraph = MetaGraph(
-            graph;  # underlying graph structure
-            label_type=Symbol,  # node name
-            vertex_data_type=AbstractNode,  # node type
-            edge_data_type=Bond,  # bond
-            graph_data=name,  # tag for the whole graph
-            # TODO add weight function and default weight
-        )
-        return new(name, metagraph)
-    end
-end
-NewBondGraph(; name=:NewBG) = NewBondGraph(DiGraph(); name=Symbol(name))
-
-name(bg::NewBondGraph) = bg.name
-size(bg::NewBondGraph) = (nv(bg.graph), ne(bg.graph))
-graph(bg::NewBondGraph) = bg.graph
-
-show(io::IO, bg::NewBondGraph) = print(io, "$(name(bg)) BondGraph$(size(bg))")
-
-# Graphs.jl interface https://juliagraphs.org/Graphs.jl/stable/ecosystem/interface/
-# FIXME arguably we don't need to do this and instead just get the graph structure from
-# the NewBG with graph(bg)
-# we only really need to access the graph for plotting and other specialised routines
-# import Graphs: edges, edgetype, has_edge, has_vertex, inneighbors, outneighbors, ne, nv, vertices, is_directed
-# edges(bg::NewBondGraph) = edges(bg.graph)
-# edgetype(bg::NewBondGraph) = edgetype(bg.graph)
-# has_edge(bg::NewBondGraph, src, dst) = has_edge(bg.graph, src, dst)
-# has_edge(bg::NewBondGraph, edge) = has_edge(bg.graph, edge)
-# has_vertex(bg::NewBondGraph, vertex) = has_vertex(bg.graph, vertex)
-# inneighbors(bg::NewBondGraph, vertex) = inneighbors(bg.graph, vertex)
-# ne(bg::NewBondGraph) = ne(bg.graph)
-# nv(bg::NewBondGraph) = nv(bg.graph)
-# outneighbors(bg::NewBondGraph, vertex) = outneighbors(bg.graph, vertex)
-# vertices(bg::NewBondGraph) = vertices(bg.graph)
-# is_directed(bg::NewBondGraph) = true  # up for discussion
-
-bg = NewBondGraph()
-
-############################################################################
-
-function add_node!(bg::NewBondGraph, comp::Component)
-    bg.graph[comp.name] = comp
-end
-
-function connect!(bg::NewBondGraph, src_comp::Component, dst_comp::Component)
-    # TODO assuming single ports, change to allow selecting a specific port
-    bg.graph[src_comp.name, dst_comp.name] = Bond(src_comp, dst_comp)
-end
-
-############################################################################
-############################################################################
-
-@variables t
-D = Differential(t)
 
 function Capacitor(; name=:capacitor)
     variables = @variables e(t), f(t), q(t)
@@ -240,48 +157,24 @@ function ZeroJunction(; name=:zero)
     return Component{EqualEffort}(; name)
 end
 
-############################################################################
 
-@named rc_model = NewBondGraph()
+############################
+# TODO determine whether a defined MTK.AbstractSystem fits the component definition
+# e.g. a storage or dissipative system is correctly described as such
+using ModelingToolkitStandardLibrary
 
-capacitor = Capacitor()
-resistor = Resistor()
-kvl = ZeroJunction()
+using ModelingToolkitStandardLibrary.Electrical
+using ModelingToolkitStandardLibrary.Blocks: Constant
 
-add_node!(rc_model, capacitor)
-add_node!(rc_model, resistor)
-add_node!(rc_model, kvl)
+typeof(__Resistor__)
+ModelingToolkit.Model
 
-connect!(rc_model, capacitor, kvl)
-connect!(rc_model, kvl, resistor)
+@named c = Capacitor()
+hierarchy(c)
+typeof(Capacitor)
+typeof(c)
 
-rc_model
-
-# Graphs.jl
-G = graph(rc_model)
-incidence_matrix(G)
-
-G[:capacitor]
-G.vertex_properties
-G.edge_data
-G.vertex_labels
-
-
-############################################################################
-using Plots, GraphRecipes
-import GraphRecipes: graphplot
-
-function graphplot(bg::NewBondGraph; kwargs...)
-    g = graph(bg)
-    graphplot(g;
-        title = name(bg),
-        names = g.vertex_labels,
-        curves = false,
-        nodeshape = :rect,
-        kwargs...
-    )
-end
-
-graphplot(rc_model)
-
-# TODO CONTINUE FROM HERE
+Capacitor.structure
+c
+equations(expand_connections(c))
+equations(c)
