+++ Project.toml
@@ -12,6 +12,7 @@ Graphs = "86223c79-3864-5bf0-83f7-82e725a168b6"
 Latexify = "23fbe1c1-3f47-55db-b15f-69d7ec21a316"
 MetaGraphsNext = "fa8bd995-216d-47f1-8a91-f3b68fbeb377"
 ModelingToolkit = "961ee093-0014-501f-94e3-6117800e7a78"
+ModelingToolkitStandardLibrary = "16a59e39-deab-5bd0-87e4-056b12336739"
 OrderedCollections = "bac558e1-5e72-5ebc-8fee-abe8a469f55d"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 RecipesBase = "3cdcf5f2-1ef4-517c-9805-6587b60abb01"
@@ -31,6 +32,7 @@ Graphs = "1"
 Latexify = "0.16"
 MetaGraphsNext = "0.7.2"
 ModelingToolkit = "9"
+ModelingToolkitStandardLibrary = "2.21.1"
 OrderedCollections = "1"
 Plots = "1.40.13"
 RecipesBase = "1"
